Split retriever unit tests into internal and external
